backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
httpx==0.28.0
beautifulsoup4==4.12.3
playwright==1.49.0
anthropic==0.40.0
python-dotenv==1.0.1
aiofiles==24.1.0
PyPDF2==3.0.1
python-docx==1.1.2
dnspython==2.7.0
pytest==8.3.4
duckduckgo_search
ddgs
